encoding: Split out the test_helper macro from maccommands tests
